Updated Collisions, Bullet Handling
Bullets are now their own prefab and have their own script.  Enemies are
now also a prefab and are instantiated as such.  Bullets now deflect
properly and are destroyed when they hit an enemy (if deflected), the
player, or go off the screen.  Player GameObject hierarchy updated to
reflect unique collisions.  See changelog/ask with questions.