Added Wave Generation, Basic Framework for Multiple Enemy Types
Enemies will now spawn in waves once there are none left on the map.
Several new methods added in EncounterManager to accomodate this.
Existing methods also heavily changed, see changelog coming soon.

Enemy prefabs added to the script can now be added to a Dictionary
(name, prefab), and when calling the CreateEnemy method, you now pass in
the name of the enemy to create.